requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
idna==3.1
lxml==4.6.2
numpy==1.20.1
pandas==1.2.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
Unidecode==1.2.0
urllib3==1.26.3
pyyaml==5.4.1